1st version
Step 1: Convert Video Images
base success
Path failure (the reason for failure is expected to be that the magnification is too large and the feature points cannot be captured. Additional tests will be conducted with other videos in the future)

Step 2: Obtain one large image through image stitching
Base failure (unable to solve the problem of long margins to the right of the image)
route failure

Step 3: Write partial code as a starting point
Completed implementation with starting point

Step 4: Implement code that combines the base and root images to represent the path
Code implementation has been roughly completed, but problems have occurred in the previous step and no progress has been made.

Code implementation method: First, in the previous step 3, the starting point image is set as a rectangle with a constant area as the starting point, and when the center of the rectangle is moved based on the gradient, the feature points are moved in the direction where they overlap the most. At this time, the maximum distance that the rectangle can move is before moving. After moving the squares, the areas of the squares overlap even slightly. After moving the squares, find the vector between the centers of the squares, and then continue to draw only those vectors on the base image to determine the movement path.
Finally we can get one big image with the movement path drawn.